refactor(amazonq): move utils functions to dedicated file (#6165)
## Problem

Lots of file utils functions in `startTransformByQ.ts` should be in
`transformFileHandler.ts`

## Solution

Move them.

---

- Treat all work as PUBLIC. Private `feature/x` branches will not be
squash-merged at release time.
- Your code changes must meet the guidelines in
[CONTRIBUTING.md](https://github.com/aws/aws-toolkit-vscode/blob/master/CONTRIBUTING.md#guidelines).

License: I confirm that my contribution is made under the terms of the
Apache 2.0 license.

---------
