+++ packages/core/src/amazonqGumby/index.ts
@@ -9,4 +9,5 @@ export { default as MessengerUtils } from './chat/controller/messenger/messenger
 export { GumbyController } from './chat/controller/controller'
 export { TabsStorage } from '../amazonq/webview/ui/storages/tabsStorage'
 export * as startTransformByQ from '../../src/codewhisperer/commands/startTransformByQ'
+export { setMaven } from '../../src/codewhisperer/service/transformByQ/transformFileHandler'
 export * from './errors'
